djangoRestAPI/requirements.txt
Django==2.2.3
gunicorn[gevent]==19.9.0
django-cors-headers==3.0.2
djangorestframework==3.10.1
packaging==19.1
drf-yasg==1.16.1
bcrypt==3.1.7
mysqlclient==1.4.4
boto3==1.9.246
django-storages==1.7.2
Pillow==5.0.0
django-filter==2.2.0